COMPMID-765 Fixed missing cast that was breaking the bare metal build
Reviewed-on: https://eu-gerrit-1.euhpc.arm.com/118086
Tested-by: Jenkins <bsgcomp@arm.com>

+++ src/runtime/NEON/functions/NEDepthwiseConvolutionLayer.cpp
@@ -158,8 +158,8 @@ void NEDepthwiseConvolutionLayer::configure(ITensor *input, const ITensor *weigh
     }
 
     // Fill borders on inputs
-    PixelValue zero_in(0);
-    PixelValue zero_w(0);
+    PixelValue zero_in(static_cast<int32_t>(0));
+    PixelValue zero_w(static_cast<int32_t>(0));
     if(_is_quantized)
     {
         zero_in = PixelValue(static_cast<int32_t>(input->info()->quantization_info().offset));
